DATAFARI-194 Synonyms etc modification behavior
Synonyms Stopwords and Protwords are now automatically uploaded to ZK
and the core is automatically reloaded.

+++ debian7/bin/zkUtils/uploadconfigzk.sh
@@ -5,9 +5,13 @@
 #
 #
 
+files=${4:-"*"};
 tmpFolder=`date +%s`_zk_config;
-mkdir -p $1/.tmp/$tmpfolder;
-cp -R $1/solr/solrcloud/$3/conf/ $1/.tmp/$tmpFolder;
-rm $1/.tmp/$tmpFolder/managed-schema;
+mkdir -p $1/.tmp/$tmpFolder;
+cp -R $1/solr/solrcloud/$3/conf/$files $1/.tmp/$tmpFolder;
+if [ -e $1/.tmp/$tmpFolder/managed-schema ]
+then
+  rm $1/.tmp/$tmpFolder/managed-schema;
+fi
 "$1/solr/server/scripts/cloud-scripts/zkcli.sh" -cmd upconfig -zkhost $2 -confdir $1/.tmp/$tmpFolder -confname $3;
 rm -R $1/.tmp/$tmpFolder;
